System representation

IntervalMDP.IntervalMarkovProcess — Type

IntervalMarkovProcess

An abstract type for interval Markov processes including IntervalMarkovChain and IntervalMarkovDecisionProcess.

IntervalMDP.num_states — Method

num_states(mp::IntervalMarkovProcess)

Return the number of states.

IntervalMDP.initial_states — Method

initial_states(mp::IntervalMarkovProcess)

Return the initial states. If the initial states are not specified, return nothing.

IntervalMDP.AllStates — Type

AllStates

A type to represent all states in a Markov process. This type is used to specify all states as the initial states.

Stationary Markov Processes

IntervalMDP.StationaryIntervalMarkovProcess — Type

StationaryIntervalMarkovProcess

An abstract type for stationary interval Markov processes including IntervalMarkovDecisionProcess.

IntervalMDP.transition_prob — Method

transition_prob(mp::StationaryIntervalMarkovProcess)

Return the interval on transition probabilities.

IntervalMDP.IntervalMarkovChain — Function

IntervalMarkovChain(transition_prob::IntervalProbabilities, initial_states::InitialStates = AllStates())

Construct an Interval Markov Chain from a square matrix pair of interval transition probabilities. The initial states are optional and if not specified, all states are assumed to be initial states. The number of states is inferred from the size of the transition probability matrix.

The returned type is an IntervalMarkovDecisionProcess with only one action per state (i.e. stateptr[j + 1] - stateptr[j] == 1 for all j). This is done to unify the interface for value iteration.

IntervalMDP.IntervalMarkovDecisionProcess — Type

IntervalMarkovDecisionProcess{
    P <: IntervalProbabilities,
    VT <: AbstractVector{Int32},
    VI <: Union{AllStates, AbstractVector}
}

A type representing (stationary) Interval Markov Decision Processes (IMDP), which are Markov Decision Processes with uncertainty in the form of intervals on the transition probabilities.

Formally, let $(S, S_0, A, \bar{P}, \underbar{P})$ be an interval Markov decision processes, where $S$ is the set of states, $S_0 \subset S$ is the set of initial states, $A$ is the set of actions, and $\bar{P} : A \to \mathbb{R}^{|S| \times |S|}$ and $\underbar{P} : A \to \mathbb{R}^{|S| \times |S|}$ are functions representing the upper and lower bound transition probability matrices prespectively for each action. Then the IntervalMarkovDecisionProcess type is defined as follows: indices 1:num_states are the states in $S$, transition_prob represents $\bar{P}$ and $\underbar{P}$, actions are implicitly defined by stateptr (e.g. if stateptr[3] == 4 and stateptr[4] == 7 then the actions available to state 3 are [4, 5, 6]), and initial_states is the set of initial states $S_0$. If no initial states are specified, then the initial states are assumed to be all states in $S$.

Fields

transition_prob::P: interval on transition probabilities where columns represent source/action pairs and rows represent target states.

stateptr::VT: pointer to the start of each source state in transition_prob (i.e. transition_prob[:, stateptr[j]:stateptr[j + 1] - 1] is the transition probability matrix for source state j) in the style of colptr for sparse matrices in CSC format.

initial_states::VI: initial states.

num_states::Int32: number of states.

Examples

transition_probs = IntervalProbabilities(;
    lower = [
        0.0 0.5 0.1 0.2 0.0
        0.1 0.3 0.2 0.3 0.0
        0.2 0.1 0.3 0.4 1.0
    ],
    upper = [
        0.5 0.7 0.6 0.6 0.0
        0.6 0.5 0.5 0.5 0.0
        0.7 0.3 0.4 0.4 1.0
    ],
)

stateptr = [1, 3, 5, 6]
initial_states = [1]

mdp = IntervalMarkovDecisionProcess(transition_probs, stateptr, initial_states)

There is also a constructor for IntervalMarkovDecisionProcess where the transition probabilities are given as a list of transition probabilities for each source state.

prob1 = IntervalProbabilities(;
    lower = [
        0.0 0.5
        0.1 0.3
        0.2 0.1
    ],
    upper = [
        0.5 0.7
        0.6 0.5
        0.7 0.3
    ],
)

prob2 = IntervalProbabilities(;
    lower = [
        0.1 0.2
        0.2 0.3
        0.3 0.4
    ],
    upper = [
        0.6 0.6
        0.5 0.5
        0.4 0.4
    ],
)

prob3 = IntervalProbabilities(;
    lower = [0.0; 0.0; 1.0],
    upper = [0.0; 0.0; 1.0]
)

transition_probs = [prob1, prob2, prob3]
initial_states = [1]

mdp = IntervalMarkovDecisionProcess(transition_probs, initial_states)

IntervalMDP.stateptr — Method

stateptr(mdp::IntervalMarkovDecisionProcess)

Return the state pointer of the Interval Markov Decision Process. The state pointer is a vector of integers where the i-th element is the index of the first element of the i-th state in the transition probability matrix. I.e. transition_prob[:, stateptr[j]:stateptr[j + 1] - 1] is the transition probability matrix for source state j.

IntervalMDP.tomarkovchain — Method

tomarkovchain(mdp::IntervalMarkovDecisionProcess, strategy::AbstractVector)

Extract an Interval Markov Chain (IMC) from an Interval Markov Decision Process under a stationary strategy. The returned type remains an IntervalMarkovDecisionProcess with only one action per state. The extracted IMC is stationary.

IntervalMDP.tomarkovchain — Method

tomarkovchain(mdp::IntervalMarkovDecisionProcess, strategy::AbstractVector{<:AbstractVector})

Extract an Interval Markov Chain (IMC) from an Interval Markov Decision Process under a time-varying strategy. The returned type is a TimeVaryingIntervalMarkovDecisionProcess with only one action per state per time-step. The extracted IMC is time-varying.

Time-varying Markov Processes

IntervalMDP.TimeVaryingIntervalMarkovProcess — Type

TimeVaryingIntervalMarkovProcess

An abstract type for time-varying interval Markov processes including TimeVaryingIntervalMarkovChain.

IntervalMDP.transition_prob — Method

transition_prob(mp::TimeVaryingIntervalMarkovProcess, t)

Return the interval on transition probabilities at time step $t$ in the range 1:time_length(mp).

IntervalMDP.time_length — Method

time_length(mp::TimeVaryingIntervalMarkovProcess)

Return the time length of the time-varying interval Markov process. Model checking for this type of process must be done over for a finite time property of equal length.

IntervalMDP.TimeVaryingIntervalMarkovChain — Function

TimeVaryingIntervalMarkovChain(transition_probs::Vector{<:IntervalProbabilities}, initial_states::InitialStates = AllStates())

Construct an Interval Markov Chain from a sequence of square matrix pairs of interval transition probabilities. The initial states are optional and if not specified, all states are assumed to be initial states. The number of states is inferred from the size of the transition probability matrix.

The returned type is an TimeVaryingIntervalMarkovDecisionProcess with only one action per state (i.e. stateptr[j + 1] - stateptr[j] == 1 for all j). This is done to unify the interface for value iteration.

IntervalMDP.TimeVaryingIntervalMarkovDecisionProcess — Type

TimeVaryingIntervalMarkovDecisionProcess{P <: IntervalProbabilities, VT <: AbstractVector{Int32}, VI <: Union{AllStates, AbstractVector}}

A type representing Time-varying Interval Markov Decision Processes (IMDP), which are Markov Decision Processes with uncertainty in the form of intervals on the transition probabilities varying over time. The time variablity must be finite.

Formally, let $(S, S_0, A, \{ \bar{P}_t \}_{t \in T}, \{ \underbar{P}_t \}_{t \in T})$ be a time-varying interval Markov chain, where $S$ is the set of states, $S_0 \subset S$ is a set of initial states, $A$ is the set of actions, and $\bar{P}_t : A \to \mathbb{R}^{|S| \times |S|}$ and $\underbar{P}_t : A \to \mathbb{R}^{|S| \times |S|}$ are the upper and lower bound transition probability matrices for time step $t \in T$ prespectively. Then the TimeVaryingIntervalMarkovDecisionProcess type is defined as follows: indices 1:num_states are the states in $S$, transition_probs[t] represents $\bar{P}_t$ and $\underbar{P}_t$ at time $t$, actions are implicitly defined by stateptr (e.g. if stateptr[3] == 4 and stateptr[4] == 7 then the actions available to state 3 are [4, 5, 6]), and initial_states is the set of initial state $S_0$. If no initial states are specified, then the initial states are assumed to be all states in $S$.

Note that for time-varying models, model checking is only enabled for finite time properties of equal length.

Fields

transition_probs::Vector{P}: interval on transition probabilities.

stateptr::VT: pointer to the start of each source state in transition_prob (i.e. transition_prob[:, stateptr[j]:stateptr[j + 1] - 1] is the transition probability matrix for source state j) in the style of colptr for sparse matrices in CSC format.

initial_states::VI: initial states.

num_states::Int32: number of states.

Examples

prob1 = IntervalProbabilities(;
    lower = [
        0.0 0.5 0.0
        0.1 0.3 0.0
        0.2 0.1 1.0
    ],
    upper = [
        0.5 0.7 0.0
        0.6 0.5 0.0
        0.7 0.3 1.0
    ],
)

prob2 = IntervalProbabilities(;
    lower = [
        0.2 0.1 0.0
        0.1 0.3 0.0
        0.0 0.5 1.0
    ],
    upper = [
        0.7 0.3 0.0
        0.6 0.5 0.0
        0.5 0.7 1.0
    ],
)

mc = TimeVaryingIntervalMarkovDecisionProcess([prob1, prob2])
# or
initial_states = [1, 2, 3]
mc = TimeVaryingIntervalMarkovDecisionProcess([prob1, prob2], initial_states)

IntervalMDP.stateptr — Method

stateptr(mdp::TimeVaryingIntervalMarkovDecisionProcess)

Return the state pointer of the Time-Varying Interval Markov Decision Process. The state pointer is a vector of integers where the i-th element is the index of the first element of the i-th state in the transition probability matrix. I.e. transition_prob[:, stateptr[j]:stateptr[j + 1] - 1] is the transition probability matrix for source state j.

IntervalMDP.tomarkovchain — Method

tomarkovchain(mdp::TimeVaryingIntervalMarkovDecisionProcess, strategy::AbstractVector{<:AbstractVector})

Extract an Interval Markov Chain (IMC) from a Time-Varying Interval Markov Decision Process under a time-varying strategy. The length of strategy must be equal to the time length of the time-varying model. The returned type is a TimeVaryingIntervalMarkovDecisionProcess with only one action per state per time-step. The extracted IMC is time-varying.

Probability representation

IntervalMDP.IntervalProbabilities — Type

IntervalProbabilities{R, VR <: AbstractVector{R}, MR <: AbstractMatrix{R}}

A matrix pair to represent the lower and upper bound transition probabilities from all source states or source/action pairs to all target states. The matrices can be Matrix{R} or SparseMatrixCSC{R}, or their CUDA equivalents. For memory efficiency, it is recommended to use sparse matrices.

The columns represent the source and the rows represent the target, as if the probability matrix was a linear transformation. Mathematically, let $P$ be the probability matrix. Then $P_{ij}$ represents the probability of transitioning from state $j$ (or with state/action pair $j$) to state $i$. Due to the column-major format of Julia, this is also a more efficient representation (in terms of cache locality).

The lower bound is explicitly stored, while the upper bound is computed from the lower bound and the gap. This choice is because it simplifies repeated probability assignment using O-maximization [1].

Fields

lower::MR: The lower bound transition probabilities from a source state or source/action pair to a target state.

gap::MR: The gap between upper and lower bound transition probabilities from a source state or source/action pair to a target state.

sum_lower::VR: The sum of lower bound transition probabilities from a source state or source/action pair to all target states.

Examples

dense_prob = IntervalProbabilities(;
    lower = [0.0 0.5; 0.1 0.3; 0.2 0.1],
    upper = [0.5 0.7; 0.6 0.5; 0.7 0.3],
)

sparse_prob = IntervalProbabilities(;
    lower = sparse_hcat(
        SparseVector(15, [4, 10], [0.1, 0.2]),
        SparseVector(15, [5, 6, 7], [0.5, 0.3, 0.1]),
    ),
    upper = sparse_hcat(
        SparseVector(15, [1, 4, 10], [0.5, 0.6, 0.7]),
        SparseVector(15, [5, 6, 7], [0.7, 0.5, 0.3]),
    ),
)

[1] M. Lahijanian, S. B. Andersson and C. Belta, "Formal Verification and Synthesis for Discrete-Time Stochastic Systems," in IEEE Transactions on Automatic Control, vol. 60, no. 8, pp. 2031-2045, Aug. 2015, doi: 10.1109/TAC.2015.2398883.

IntervalMDP.lower — Method

lower(p::IntervalProbabilities)

Return the lower bound transition probabilities from a source state or source/action pair to a target state.

IntervalMDP.upper — Method

upper(p::IntervalProbabilities)

Return the upper bound transition probabilities from a source state or source/action pair to a target state.

Note

It is not recommended to use this function for the hot loop of O-maximization. Because the IntervalProbabilities stores the lower and gap transition probabilities, fetching the upper bound requires allocation and computation.

IntervalMDP.gap — Method

gap(p::IntervalProbabilities)

Return the gap between upper and lower bound transition probabilities from a source state or source/action pair to a target state.

IntervalMDP.sum_lower — Method

sum_lower(p::IntervalProbabilities)

Return the sum of lower bound transition probabilities from a source state or source/action pair to all target states. This is useful in efficiently implementing O-maximization, where we start with a lower bound probability assignment and iteratively, according to the ordering, adding the gap until the sum of probabilities is 1.

IntervalMDP.num_source — Method

num_source(p::IntervalProbabilities)

Return the number of source states or source/action pairs.

IntervalMDP.num_target — Method

num_target(p::IntervalProbabilities)

Return the number of target states.

IntervalMDP.axes_source — Method

axes_source(p::IntervalProbabilities)

Return the valid range of indices for the source states or source/action pairs.