System representation

IntervalMDP.IntervalMarkovProcess — Type

IntervalMarkovProcess

An abstract type for interval Markov processes including IntervalMarkovChain and IntervalMarkovDecisionProcess.

Markov chain

IntervalMDP.IntervalMarkovChain — Type

IntervalMarkovChain{P <: IntervalProbabilities, T <: Integer, VT <: AbstractVector{T}}

A type representing Interval Markov Chains (IMC), which are Markov Chains with uncertainty in the form of intervals on the transition probabilities.

Formally, let $(S, S_0, \bar{P}, \underbar{P})$ be an interval Markov chain, where $S$ is the set of states, $S_0 \subset S$ is a set of initial states, and $\bar{P} : \mathbb{R}^{|S| \times |S|}$ and $\underbar{P} : \mathbb{R}^{|S| \times |S|}$ are the upper and lower bound transition probability matrices prespectively. Then the IntervalMarkovChain type is defined as follows: indices 1:num_states are the states in $S$, transition_prob represents $\bar{P}$ and $\underbar{P}$, and initial_states is the set of initial state $S_0$. If no initial states are specified, then the initial states are assumed to be all states in $S$.

Fields

transition_prob::P: interval on transition probabilities.

initial_states::VT: initial states.

num_states::T: number of states.

Examples

prob = IntervalProbabilities(;
    lower = [
        0.0 0.5 0.0
        0.1 0.3 0.0
        0.2 0.1 1.0
    ],
    upper = [
        0.5 0.7 0.0
        0.6 0.5 0.0
        0.7 0.3 1.0
    ],
)

mc = IntervalMarkovChain(prob)
# or
initial_states = [1, 2, 3]
mc = IntervalMarkovChain(prob, initial_states)

IntervalMDP.transition_prob — Method

transition_prob(s::IntervalMarkovChain)

Return the interval on transition probabilities.

IntervalMDP.num_states — Method

num_states(s::IntervalMarkovChain)

Return the number of states.

IntervalMDP.initial_states — Method

initial_states(s::IntervalMarkovChain)

Return the initial states.

Markov decision process

IntervalMDP.IntervalMarkovDecisionProcess — Type

IntervalMarkovDecisionProcess{
    P <: IntervalProbabilities,
    T <: Integer,
    VT <: AbstractVector{T},
    VI <: AbstractVector{T},
    VA <: AbstractVector,
}

A type representing Interval Markov Decision Processes (IMDP), which are Markov Decision Processes with uncertainty in the form of intervals on the transition probabilities.

Formally, let $(S, S_0, A, \bar{P}, \underbar{P})$ be an interval Markov decision processes, where $S$ is the set of states, $S_0 \subset S$ is the set of initial states, $A$ is the set of actions, and $\bar{P} : A \to \mathbb{R}^{|S| \times |S|}$ and $\underbar{P} : A \to \mathbb{R}^{|S| \times |S|}$ are functions representing the upper and lower bound transition probability matrices prespectively for each action. Then the IntervalMarkovDecisionProcess type is defined as follows: indices 1:num_states are the states in $S$, transition_prob represents $\bar{P}$ and $\underbar{P}$, action_vals contains the actions available in each state, and initial_states is the set of initial states $S_0$. If no initial states are specified, then the initial states are assumed to be all states in $S$.

Fields

transition_prob::P: interval on transition probabilities where columns represent source/action pairs and rows represent target states.

stateptr::VT: pointer to the start of each source state in transition_prob (i.e. transition_prob[:, stateptr[j]:stateptr[j + 1] - 1] is the transition probability matrix for source state j) in the style of colptr for sparse matrices in CSC format.

action_vals::VA: actions available in each state. Can be any eltype.

initial_states::VI: initial states.

num_states::T: number of states.

Examples

transition_probs = IntervalProbabilities(;
    lower = [
        0.0 0.5 0.1 0.2 0.0
        0.1 0.3 0.2 0.3 0.0
        0.2 0.1 0.3 0.4 1.0
    ],
    upper = [
        0.5 0.7 0.6 0.6 0.0
        0.6 0.5 0.5 0.5 0.0
        0.7 0.3 0.4 0.4 1.0
    ],
)

stateptr = [1, 3, 5, 6]
actions = ["a1", "a2", "a1", "a2", "sinking"]
initial_states = [1]

mdp = IntervalMarkovDecisionProcess(transition_probs, stateptr, actions, initial_states)

There is also a constructor for IntervalMarkovDecisionProcess where the transition probabilities are given as a list of mappings from actions to transition probabilities for each source state.

prob1 = IntervalProbabilities(;
    lower = [
        0.0 0.5
        0.1 0.3
        0.2 0.1
    ],
    upper = [
        0.5 0.7
        0.6 0.5
        0.7 0.3
    ],
)

prob2 = IntervalProbabilities(;
    lower = [
        0.1 0.2
        0.2 0.3
        0.3 0.4
    ],
    upper = [
        0.6 0.6
        0.5 0.5
        0.4 0.4
    ],
)

prob3 = IntervalProbabilities(;
    lower = [0.0; 0.0; 1.0],
    upper = [0.0; 0.0; 1.0]
)

transition_probs = [["a1", "a2"] => prob1, ["a1", "a2"] => prob2, ["sinking"] => prob3]
initial_states = [1]

mdp = IntervalMarkovDecisionProcess(transition_probs, initial_states)

IntervalMDP.transition_prob — Method

transition_prob(s::IntervalMarkovDecisionProcess)

Return the interval on transition probabilities.

IntervalMDP.num_states — Method

num_states(s::IntervalMarkovDecisionProcess)

Return the number of states.

IntervalMDP.initial_states — Method

initial_states(s::IntervalMarkovDecisionProcess)

Return the initial states.

IntervalMDP.actions — Method

actions(s::IntervalMarkovDecisionProcess)

Return a vector of actions (choices in PRISM terminology).

IntervalMDP.num_choices — Method

num_choices(s::IntervalMarkovDecisionProcess)

Return the sum of the number of actions available in each state $\sum_{j} \mathrm{num\_actions}(s_j)$.

Probability representation

IntervalMDP.IntervalProbabilities — Type

IntervalProbabilities{R, VR <: AbstractVector{R}, MR <: AbstractMatrix{R}}

A matrix pair to represent the lower and upper bound transition probabilities from a source state or source/action pair to a target state. The matrices can be Matrix{R} or SparseMatrixCSC{R}, or their CUDA equivalents. For memory efficiency, it is recommended to use sparse matrices.

The columns represent the source and the rows represent the target, as if the probability matrix was a linear transformation. Mathematically, let $P$ be the probability matrix. Then $P_{ij}$ represents the probability of transitioning from state $j$ (or with state/action pair $j$) to state $i$. Due to the column-major format of Julia, this is also a more efficient representation (in terms of cache locality).

The lower bound is explicitly stored, while the upper bound is computed from the lower bound and the gap. This choice is because it simplifies repeated probability assignment using O-maximization [1].

Fields

lower::MR: The lower bound transition probabilities from a source state or source/action pair to a target state.

gap::MR: The gap between upper and lower bound transition probabilities from a source state or source/action pair to a target state.

sum_lower::VR: The sum of lower bound transition probabilities from a source state or source/action pair to all target states.

Examples

dense_prob = IntervalProbabilities(;
    lower = [0.0 0.5; 0.1 0.3; 0.2 0.1],
    upper = [0.5 0.7; 0.6 0.5; 0.7 0.3],
)

sparse_prob = IntervalProbabilities(;
    lower = sparse_hcat(
        SparseVector(15, [4, 10], [0.1, 0.2]),
        SparseVector(15, [5, 6, 7], [0.5, 0.3, 0.1]),
    ),
    upper = sparse_hcat(
        SparseVector(15, [1, 4, 10], [0.5, 0.6, 0.7]),
        SparseVector(15, [5, 6, 7], [0.7, 0.5, 0.3]),
    ),
)

[1] M. Lahijanian, S. B. Andersson and C. Belta, "Formal Verification and Synthesis for Discrete-Time Stochastic Systems," in IEEE Transactions on Automatic Control, vol. 60, no. 8, pp. 2031-2045, Aug. 2015, doi: 10.1109/TAC.2015.2398883.

IntervalMDP.lower — Method

lower(s::IntervalProbabilities)

Return the lower bound transition probabilities from a source state or source/action pair to a target state.

IntervalMDP.upper — Method

upper(s::IntervalProbabilities)

Return the upper bound transition probabilities from a source state or source/action pair to a target state.

Note

It is not recommended to use this function for the hot loop of O-maximization, because it is not just an accessor and requires allocation and computation.

IntervalMDP.gap — Method

gap(s::IntervalProbabilities)

Return the gap between upper and lower bound transition probabilities from a source state or source/action pair to a target state.

IntervalMDP.sum_lower — Method

sum_lower(s::IntervalProbabilities)

Return the sum of lower bound transition probabilities from a source state or source/action pair to all target states. This is useful in efficiently implementing O-maximization, where we start with a lower bound probability assignment and iteratively, according to the ordering, adding the gap until the sum of probabilities is 1.

IntervalMDP.num_source — Method

num_source(s::IntervalProbabilities)

Return the number of source states or source/action pairs.

IntervalMDP.num_target — Method

num_target(s::IntervalProbabilities)

Return the number of target states.