Value iteration

IntervalMDP.value_iteration — Function

value_iteration(problem::Problem{<:IntervalMarkovChain, <:Specification})

Solve optimistic/pessimistic specification problems using value iteration for interval Markov chain.

Examples

prob = IntervalProbabilities(;
    lower = [
        0.0 0.5 0.0
        0.1 0.3 0.0
        0.2 0.1 1.0
    ],
    upper = [
        0.5 0.7 0.0
        0.6 0.5 0.0
        0.7 0.3 1.0
    ],
)

mc = IntervalMarkovChain(prob, 1)

terminal_states = [3]
time_horizon = 10
prop = FiniteTimeReachability(terminal_states, time_horizon)
spec = Specification(prop, Pessimistic)
problem = Problem(mc, spec)
V, k, residual = value_iteration(problem)

value_iteration(problem::Problem{<:IntervalMarkovDecisionProcess, <:Specification})

Solve minimizes/mazimizes optimistic/pessimistic specification problems using value iteration for interval Markov decision processes.

Examples

prob1 = IntervalProbabilities(;
    lower = [
        0.0 0.5
        0.1 0.3
        0.2 0.1
    ],
    upper = [
        0.5 0.7
        0.6 0.5
        0.7 0.3
    ],
)

prob2 = IntervalProbabilities(;
    lower = [
        0.1 0.2
        0.2 0.3
        0.3 0.4
    ],
    upper = [
        0.6 0.6
        0.5 0.5
        0.4 0.4
    ],
)

prob3 = IntervalProbabilities(;
    lower = [0.0; 0.0; 1.0],
    upper = [0.0; 0.0; 1.0]
)

transition_probs = [["a1", "a2"] => prob1, ["a1", "a2"] => prob2, ["sinking"] => prob3]
initial_state = 1
mdp = IntervalMarkovDecisionProcess(transition_probs, initial_state)

terminal_states = [3]
time_horizon = 10
prop = FiniteTimeReachability(terminal_states, time_horizon)
spec = Specification(prop, Pessimistic, Maximize)
problem = Problem(mdp, spec)
V, k, residual = value_iteration(problem)

O-maximization

IntervalMDP.ominmax — Function

ominmax(prob, V; max = true)

Compute probability assignment within the interval probabilities prob that upper or lower bounds the expectation of the value function V via O-maximization [1]. Whether the expectation is maximized or minimized is determined by the max keyword argument. That is, if max == true then an upper bound is computed and if max == false then a lower bound is computed.

Examples

prob = IntervalProbabilities(;
    lower = sparse_hcat(
        SparseVector(15, [4, 10], [0.1, 0.2]),
        SparseVector(15, [5, 6, 7], [0.5, 0.3, 0.1]),
    ),
    upper = sparse_hcat(
        SparseVector(15, [1, 4, 10], [0.5, 0.6, 0.7]),
        SparseVector(15, [5, 6, 7], [0.7, 0.5, 0.3]),
    ),
)

V = collect(1:15)
p = ominmax(prob, V; max = true)

Note

This function will construct a workspace object for the ordering and an output vector. For a hot-loop, it is more efficient to use ominmax! and pass in pre-allocated objects. See construct_ordering for how to pre-allocate the workspace.

[1] M. Lahijanian, S. B. Andersson and C. Belta, "Formal Verification and Synthesis for Discrete-Time Stochastic Systems," in IEEE Transactions on Automatic Control, vol. 60, no. 8, pp. 2031-2045, Aug. 2015, doi: 10.1109/TAC.2015.2398883.

IntervalMDP.ominmax! — Function

ominmax!(ordering, p, prob, V; max = true)

Compute in-place the probability assignment within the interval probabilities prob that upper or lower bounds the expectation of the value function V via O-maximization [1]. Whether the expectation is maximized or minimized is determined by the max keyword argument. That is, if max == true then an upper bound is computed and if max == false then a lower bound is computed.

The output is constructed in the input vector p and returned. The ordering workspace object is also modified.

Examples

prob = IntervalProbabilities(;
    lower = sparse_hcat(
        SparseVector(15, [4, 10], [0.1, 0.2]),
        SparseVector(15, [5, 6, 7], [0.5, 0.3, 0.1]),
    ),
    upper = sparse_hcat(
        SparseVector(15, [1, 4, 10], [0.5, 0.6, 0.7]),
        SparseVector(15, [5, 6, 7], [0.7, 0.5, 0.3]),
    ),
)

V = collect(1:15)
ordering = construct_ordering(prob)
p = deepcopy(gap(p))

p = ominmax!(ordering, p, prob, V; max = true)

[1] M. Lahijanian, S. B. Andersson and C. Belta, "Formal Verification and Synthesis for Discrete-Time Stochastic Systems," in IEEE Transactions on Automatic Control, vol. 60, no. 8, pp. 2031-2045, Aug. 2015, doi: 10.1109/TAC.2015.2398883.

IntervalMDP.partial_ominmax — Function

partial_ominmax(prob, V, indices; max = true)

Perform O-maximization on a subset of source states or source/action pairs according to indices. This corresponds to the columns in prob. See ominmax for more details on what O-maximization is.

Note

This function will construct a workspace object for the ordering and an output vector. For a hot-loop, it is more efficient to use ominmax! and pass in pre-allocated objects. See construct_ordering for how to pre-allocate the workspace.

IntervalMDP.partial_ominmax! — Function

partial_ominmax!(ordering, p, prob, V, indices; max = true)

Perform O-maximization in-place on a subset of source states or source/action pairs according to indices. This corresponds to the columns in prob. See ominmax for more details on what O-maximization is.

IntervalMDP.construct_ordering — Function

construct_ordering(p::AbstractMatrix)

Construct a workspace for constructing and storing orderings of states, given a value function. If O-maximization is used in a hot-loop, it is more efficient to use this function to preallocate the workspace and reuse across iterations.

An alternative constructor is construct_ordering(prob::IntervalProbabilities).